Move the compiler closer to dialects & databases that it uses & is used in
